Dialogue System Testing
Got a node based dialogue system working with a script to disable player movement inputs during the dialogue. Next step is to add an npc that can be interacted with to activate the dialogue box.
Also I fixed an oversight with the player controller that was causing the jump animation to not play.