NPC With Cutscene Logic
Added an NPC with a dialogue tree which is activated when the player presses interact within the npc's trigger collider. The dialogue tree will end with a function being called that will play a cutscene.

I tried like 6 different 3d models for the character, and finally found one that had correct bones for a humanoid rig. And then I had to manually connect the textures to the materials on the model.

I also had to change the material surface type on a few parts of the model that were not rendering properly.